#177 moved password module file in helpers dir
