+++ src/controllers/helpers/passwordModule.js
