+++ src/controllers/signup.js
@@ -1,4 +1,4 @@
-const { hashPassword } = require('./passwordModule');
+const { hashPassword } = require('./helpers/passwordModule');
 const { getUser, addUser } = require('../model/user_queries');
 const { validateSignUp } = require('./validate');
